Multiclass linear SVM

+++ cpp/model/Model.hpp
@@ -8,7 +8,7 @@
 class Model {
     virtual void _train(const std::vector<Data> &training, const std::vector<Data> &validation) = 0;
     virtual void _save(const std::string &filename) const = 0;
-    [[nodiscard]] virtual short _call(const Data &input) const = 0;
+    [[nodiscard]] virtual int _call(const Data &input) const = 0;
     [[nodiscard]] virtual double _error(const std::vector<Data> &test) const;
 public:
     inline void train(const std::vector<Data> &training, const std::vector<Data> &validation = std::vector<Data>()) {
@@ -20,7 +20,7 @@ public:
     [[nodiscard]] inline double error(const std::vector<Data> &test) const {
         return _error(test);
     }
-    inline short operator()(const Data &input) const {
+    inline int operator()(const Data &input) const {
         return _call(input);
     }
     virtual ~Model() = default;
